fix: color and display button name
